+++ src/BrowserModuleResolver.ts
@@ -20,7 +20,7 @@ import * as meriyahPlugin from "prettier/parser-meriyah";
 import * as typescriptPlugin from "prettier/parser-typescript";
 import * as yamlPlugin from "prettier/parser-yaml";
 
-// commente out as the cod uses `new Function("return this") which
+// commented out as the cod uses `new Function("return this") which
 // is not allowed in the VS Code extension host as it enforces
 // the Trusted Types Content Security Policy
 //import * as flowPlugin from "prettier/parser-flow";